增加token过期和刷新token 实现用户唯一

+++ SpringSecurityQuickStarter/src/main/java/com/lsy/handler/JwtLogoutSuccessHandler.java
@@ -1,8 +1,10 @@
 package com.lsy.handler;
 
 import com.alibaba.fastjson2.JSON;
+import com.lsy.entity.LoginUser;
 import com.lsy.utils.*;
 import io.jsonwebtoken.Claims;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
 import org.springframework.stereotype.Component;
@@ -12,6 +14,8 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.time.LocalDateTime;
+import java.util.Objects;
 
 /**
  * @ClassName JwtLogoutSuccessHandler
@@ -21,6 +25,7 @@ import java.io.IOException;
  * @Version 1.0
  */
 @Component
+@Slf4j
 public class JwtLogoutSuccessHandler implements LogoutSuccessHandler {
 
     @Resource
@@ -41,8 +46,14 @@ public class JwtLogoutSuccessHandler implements LogoutSuccessHandler {
             e.printStackTrace();
         }
         String userId = claims.getSubject();
-        //redis中移除用户
-        redisCache.deleteObject("login:" + userId);
+        LoginUser loginUser = redisCache.getCacheObject("login:" + userId);
+        if(Objects.nonNull(loginUser)) {
+            log.info(String.format("IP %s，用户 %s， 于 %s 退出系统。",
+                    request.getRemoteHost(), loginUser.getSysUser().getUserName(), LocalDateTime.now()));
+            //redis中移除用户
+            redisCache.deleteObject("login:" + userId);
+        }
+
         String msg = StringUtils.format("用户【" + userId + "】注销成功！");
         ServletUtils.renderString(response, JSON.toJSONString(AjaxResult.success(msg)));
     }
